backend/requirements.txt
# web framework
fastapi==0.116.1
# ASGI server with WebSocket support
uvicorn[standard]==0.22.0        # the “[standard]” extra pulls in websockets/wsproto etc.
# if you prefer to pin them yourself instead of using the extra:
# websockets==11.0.3
# wsproto==1.2.0
gunicorn

# pydantic & typing
pydantic==2.11.7
pydantic[email]
email-validator>=2.1.0
pydantic-settings>=1.0.0
typing-extensions==4.14.1
annotated-types==0.7.0
typing-inspection==0.4.1

# async plumbing
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
idna==3.10
click==8.2.1

# date handling
python-dateutil==2.9.0.post0

prometheus_client
asyncpg
sqlalchemy[asyncio]
python-jose[cryptography]>=3.3.0
python-dotenv
python-multipart

passlib[bcrypt]
bcrypt<4.1.0

alembic
rich
psycopg2-binary
boto3
aiohttp
# your SQLite “vault” needs no extra deps

# (optional) if you ever log to JSON, you might pull in python-json-logger,
# but it’s not required for your current code.